use existing timestamp formatters for comparison chart

+++ src/chart/scale.rs
@@ -276,7 +276,7 @@ impl AxisLabelsX<'_> {
                 if let Some(timestamp) = interval_keys.get(array_index) {
                     let text_content = self
                         .timezone
-                        .format_crosshair_timestamp(*timestamp as i64, interval.0.into());
+                        .format_crosshair_timestamp(*timestamp, interval.0.into());
 
                     return Some(AxisLabel::new_x(
                         snap_x,
@@ -314,7 +314,7 @@ impl AxisLabelsX<'_> {
 
                 let text_content = self
                     .timezone
-                    .format_crosshair_timestamp(rounded_timestamp as i64, interval);
+                    .format_crosshair_timestamp(rounded_timestamp, interval);
 
                 return Some(AxisLabel::new_x(
                     snap_x as f32,
